requirements.txt
Django==1.11
dj-database-url==0.4.1
gunicorn==19.6.0
django-grappelli==2.10.1
whitenoise==3.3.0
psycopg2==2.6.2
django-nested-inline==0.3.6